Refactor express apis for testing
